Implementando caso de uso 03(parcial)

+++ src/main/java/dao/CursoDao.java
@@ -1,5 +1,6 @@
 package dao;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import dominio.Curso;
@@ -10,7 +11,7 @@ public interface CursoDao {
 	public void excluir(Curso x);
 	public Curso buscar(int cod);
 	public List<Curso> buscarTodos();
-	public List<Curso> buscarPorNome(String nome,Double precoMinimo,Double precoMaximo);
+	public List<Curso> buscarPorNome(String nome,BigDecimal precoMinimo,BigDecimal precoMaximo);
 	public List<Curso> buscarCursosOrderNome();
 	public Curso buscarNomeExato(String nome);
 	public List<Curso> buscarPorTrecho(String trecho);
