+++ src/main/java/servico/CursoServico.java
@@ -1,5 +1,6 @@
 package servico;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -102,7 +103,7 @@ public class CursoServico {
 		return dao.buscarTodos();
 	}
 
-	public List<Curso> buscarPorNome(String nome,Double precoMinimo,Double precoMaximo){
+	public List<Curso> buscarPorNome(String nome,BigDecimal precoMinimo,BigDecimal precoMaximo){
 		return dao.buscarPorNome(nome,precoMinimo,precoMaximo);
 	}
